+++ helpers.go
@@ -1,14 +1,13 @@
 package fmc
 
-import "crypto/rand"
+import "math/rand/v2"
 
 // generate random string
 func generateRequestID(length int) string {
 	const charset = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
 	b := make([]byte, length)
-	rand.Read(b)
 	for i := range b {
-		b[i] = charset[b[i]%byte(len(charset))]
+		b[i] = charset[rand.IntN(len(charset))]
 	}
 	return string(b)
 }